Modulo 2 - Exercicio 15 resolvido

+++ modulo2/ex05/swap.h
@@ -1,4 +1,4 @@
 #ifndef SWAP_H
 #define SWAP_H
-	short swap();
+	int swap();
 #endif
